app0-app1/requirements.txt
hopeit.engine[web,cli,fs-storage,redis-streams,apps-visualizer,config-manager,apps-client]>=0.15.0,<0.16.0
app0.admin>=0.1.5
app0.platform.auth>=0.1.5
motor>=2.5.1
aioboto3>=9.3.1
jinja2>=3.0.3
